TurnOnBoat

+++ Handle/Unit/Survivor.h
@@ -18,7 +18,7 @@ class Survivor : public Unit { //survivor is the derived class of unit
         explicit Survivor(Map *map, double x, double y); //player constructor
         static const int base_max_health; //base maximum health of player
         static const float base_collision_radius; //base collision radius of player
-        static const int base_pickup_radius; // base searching area for item of player
+        static const float base_pickup_radius; // base searching area for item of player
         static const double base_max_speed; //base maximum movement speed 
         
         static const int base_damage = 0; //base melee damage
